mobile reader: extract useReaderBars hook
Reader screen is 1419 lines; immersive-mode bars (state + animated value + 4 handlers + cleanup effect) are self-contained → moved to src/hooks/useReaderBars.ts. -43 lines net in the screen.
